requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv==1.0.0
Pillow==10.1.0
mysqlclient==2.2.0
gunicorn==21.2.0
